+++ scilab/modules/graphics/src/c/getHandleProperty/set_arc_drawing_method_property.c
@@ -25,7 +25,7 @@
 #include "SetPropertyStatus.h"
 
 /*------------------------------------------------------------------------*/
-int set_arc_drawing_method_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_arc_drawing_method_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
 
   if ( sciGetEntityType(pobj) != SCI_ARC && sciGetEntityType(pobj) != SCI_SUBWIN )
