+++ scilab/modules/graphics/src/c/getHandleProperty/set_callbackmevent_property.c
@@ -28,7 +28,7 @@
 #include "SetPropertyStatus.h"
 
 /*------------------------------------------------------------------------*/
-int set_callbackmevent_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_callbackmevent_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
   if ( !isParameterDoubleMatrix( valueType ) )
   {
