+++ scilab/modules/graphics/src/c/getHandleProperty/set_children_property.c
@@ -26,7 +26,7 @@
 #include "SetPropertyStatus.h"
 
 /*------------------------------------------------------------------------*/
-int set_children_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_children_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
   sciprint("Children property can not be modified directly.\n") ;
   return SET_PROPERTY_ERROR ;
