+++ scilab/modules/graphics/src/c/getHandleProperty/set_clip_state_property.c
@@ -27,7 +27,7 @@
 #include "SetPropertyStatus.h"
 
 /*------------------------------------------------------------------------*/
-int set_clip_state_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_clip_state_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
 
   if ( !isParameterStringMatrix( valueType ) )
