+++ scilab/modules/graphics/src/c/getHandleProperty/set_links_property.c
@@ -28,7 +28,7 @@
 #include "localization.h"
 
 /*------------------------------------------------------------------------*/
-int set_links_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_links_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
   int nblegends=pLEGEND_FEATURE(pobj)->nblegends;
   int i;
