+++ scilab/modules/graphics/src/c/getHandleProperty/set_pixmap_property.c
@@ -26,7 +26,7 @@
 #include "GetProperty.h"
 
 /*------------------------------------------------------------------------*/
-int set_pixmap_property( sciPointObj * pobj, int stackPointer, int valueType, int nbRow, int nbCol )
+int set_pixmap_property( sciPointObj * pobj, size_t stackPointer, int valueType, int nbRow, int nbCol )
 {
   if ( sciGetEntityType (pobj) != SCI_FIGURE )
   {
